modelos criados e configurados
