+++ functions/src/commands/modelo.ts
@@ -1,6 +1,7 @@
 // src/commands/modelo.ts
 import { Context, Telegraf, Markup } from "telegraf";
 import { listModelsFromFolder } from "../services/google";
+import { getPreviewTitle } from "../utils/utils";
 
 export function getName() {
   return "/modelo";
@@ -12,13 +13,6 @@ export function getHelp() {
 
 export function getDescription() {
   return "📄 Criar documento a partir de modelo disponível.";
-}
-
-function formatDate(date: Date): string {
-  const yyyy = date.getFullYear();
-  const mm = ("0" + (date.getMonth() + 1)).slice(-2);
-  const dd = ("0" + date.getDate()).slice(-2);
-  return `${yyyy}.${mm}.${dd}`;
 }
 
 export async function register(bot: Telegraf) {
@@ -34,11 +28,10 @@ export async function register(bot: Telegraf) {
       );
       return;
     }
+    // Utiliza apenas o título informado, sem data, deixando para o callback adicionar a data atual.
     const newTitleProvided = args.join(" ");
-    const todayFormatted = formatDate(new Date());
-    const finalTitle = `${todayFormatted} - ${newTitleProvided}`;
-    // Monta a mensagem que será exibida
-    const textMessage = `Qual o modelo de documento você quer clonar?\nTítulo do documento: ${finalTitle}`;
+    // Monta a mensagem exibida para o usuário
+    const textMessage = `Qual o modelo de documento você quer clonar?\nTítulo do documento: ${newTitleProvided}`;
     try {
       const models = await listModelsFromFolder(
         "1xYWUMDsamike4q3_CrHAZNKNsq6gsUUb"
@@ -47,9 +40,15 @@ export async function register(bot: Telegraf) {
         await ctx.reply("Nenhum modelo disponível.");
         return;
       }
-      // Gera os botões – a callback_data inclui apenas o templateId
+
+      // Gera os botões para cada modelo; cada callback_data tem o formato:model.name modelo_<templateId>_<newTitleProvidedEncoded>
       const buttons = models.map((model) => {
-        return [{ text: model.name, callback_data: `modelo_${model.id}` }];
+        return [
+          {
+            text: getPreviewTitle(model.name, newTitleProvided),
+            callback_data: `modelo_${model.id}`,
+          },
+        ];
       });
       await ctx.reply(textMessage, Markup.inlineKeyboard(buttons));
     } catch (error) {
