+++ functions/src/utils/utils.ts
@@ -1,5 +1,30 @@
 import { PaymentRequest } from "../config/types";
 
+export function getPreviewTitle(
+  name: string,
+  finalTitleFromMsg: string
+): string {
+  const modelName = name.replace("[modelo]", "").trim();
+  // Supomos que o nome do modelo tem o formato: "Tipo de documento - 2025.00.00 - Alguma coisa"
+  const parts = modelName.split(" - ");
+  if (parts.length < 3) {
+    throw new Error("Formato de nome de modelo inválido.");
+  }
+  const type = parts[0]; // Exemplo: "Ata", "Requerimento", etc.
+  const currentDate = formatDate(new Date());
+  // Constrói o novo título: "Tipo de documento - YYYY.MM.DD - Título do documento"
+  const newTitle = `${type} - ${currentDate} - ${finalTitleFromMsg}`;
+  return newTitle;
+}
+
+// Função auxiliar para formatar data no formato AAAA.MM.DD
+export function formatDate(date: Date): string {
+  const yyyy = date.getFullYear();
+  const mm = ("0" + (date.getMonth() + 1)).slice(-2);
+  const dd = ("0" + date.getDate()).slice(-2);
+  return `${yyyy}.${mm}.${dd}`;
+}
+
 export function excerptFromRequest(
   request: PaymentRequest,
   title?: string | undefined
